DSi Compatible Build
Altered memory.s to allow compiling on devkitarm r46. Works on DSi.
However due to the memory constraints of the original setup larger roms
will hang on load. (Roms as large as Mario 3 seem to load though)

Note this build is intended for DSi/TWL mode ONLY. It is not recommended
for NTR mode use!

This build I made to allow operation on DSi/3DS in TWL mode so that roms
can be loaded off SD card instead of a slot-1 device.

The "disk" folder seems to be a custom replacement to fat libraries.
This wasn't compatible with DSi SD appearently, so I've removed it and
re-enabled standard fat drivers. SD access works now. :D

Note that anything wifi related likely will not work since TWL mode wifi
does not have homebrew libraries available for it yet. Use the original
NTR builds for that. (but note they must be run from a Slot-1 Flashcart
in that case)